requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.12

# Google Gemini API
google-genai==1.0.0

# Async support
aiofiles==24.1.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
typing-extensions==4.12.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0